Use a non-persistent cookie store for the web view and copy cookies to/from shared NSHTTPCookieStorage
By default WKWebView uses a persistent cookie store. This cookie store isn't
synced (at least not predictably) with the shared NSHTTPCookieStorage. It is
also unreliable to iterate the cookies in this default cookie store.

To accomplish the desired synchornization we now configure the WKWebView with
a non-persistent cookie store. LiveAuthDialog then copies relevant cookies
from the shared NSHTTPCookieStorage to the WKWebView's cookie store before it
appears. When the dialog disappears it copies any cookies back to the shared
NSHTTPCookieStorage.

As it's only possible to configure the cookie store to use by WKWebView
programmatically, the LiveAuthDialog.xib file has been deleted and we now set
up the LiveAuthDialog UI entirely in code.